refactor(node): restructure to follow n8n best practices
- Consolidate separate properties/functions files into organized index.ts modules
- Adopt official n8n starter template structure with proper resource separation
- Improve module organization with clean category-based operation dispatch
